+++ library/src/main/java/com/github/conhea/floatingactionbutton/AddFloatingActionButton.java
@@ -8,9 +8,10 @@ import android.graphics.Paint.Style;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.ShapeDrawable;
 import android.graphics.drawable.shapes.Shape;
-import android.support.annotation.ColorRes;
-import android.support.annotation.DrawableRes;
 import android.util.AttributeSet;
+
+import androidx.annotation.ColorRes;
+import androidx.annotation.DrawableRes;
 
 public class AddFloatingActionButton extends FloatingActionButton {
     int mPlusColor;