+++ library/src/main/java/com/github/conhea/floatingactionbutton/FloatingActionButtonSnackbarBehavior.java
@@ -1,13 +1,15 @@
 package com.github.conhea.floatingactionbutton;
 
 import android.content.Context;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.Snackbar;
-import android.support.v4.view.animation.FastOutSlowInInterpolator;
 import android.util.AttributeSet;
 import android.view.View;
 
+import com.google.android.material.snackbar.Snackbar;
+
 import java.util.List;
+
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
+import androidx.interpolator.view.animation.FastOutSlowInInterpolator;
 
 
 /**